+++ tests/unit/visualize/view/tField.m
@@ -1,23 +1,6 @@
 classdef tField < MAGViewTestCase
 % TFIELD Unit tests for "mag.graphics.view.Field" class.
 
-    methods (TestClassSetup)
-
-        % Disable visibility for figures while testing.
-        function disableFigureVisibility(testCase)
-
-            currentValue = get(groot(), "DefaultFigureVisible");
-            testCase.addTeardown(@() set(groot(), DefaultFigureVisible = currentValue));
-
-            set(groot(), DefaultFigureVisible = "off");
-        end
-
-        % Close all figures opened by test.
-        function closeTestFigures(testCase)
-            testCase.applyFixture(mag.test.fixture.CleanupFigures());
-        end
-    end
-
     methods (Test)
 
         % Test that science view is generated correctly.